requirements.txt
asn1crypto==0.22.0
bleach==1.5.0
cffi==1.10.0
cryptography==1.8.1
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
fire==0.1.2
graphviz==0.8
h5py==2.7.0
html5lib==0.9999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
Keras==1.2.2
Mako==1.0.6
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mxnet-cu80==0.11.0
nb-conda-kernels==2.1.0
nbconvert==5.2.1
nbformat==4.4.0
nose==1.3.7
notebook==5.0.0
numpy==1.13.1
olefile==0.44
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
patsy==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
pycosat==0.6.2
pycparser==2.17
Pygments==2.2.0
pygpu==0.6.9
pyOpenSSL==17.0.0
pyparsing==2.1.4
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.14.2
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
six==1.10.0
statsmodels==0.8.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.5
terminado==0.6
testpath==0.3
Theano==0.9.0
torch==0.2.0.post4
torchvision==0.1.9
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.2
